Clean + Fix Datagen + Add Pale Hanging Moss

+++ src/main/java/com/natsu/backport/common/particles/CherryParticle.java
@@ -48,13 +48,13 @@ public class CherryParticle extends TextureSheetParticle {
         }
 
         if (!this.removed) {
-            float f = (float)(300 - this.lifetime);
+            float f = 300 - this.lifetime;
             float f1 = Math.min(f / 300.0F, 1.0F);
-            double d0 = Math.cos(Math.toRadians((double)(this.particleRandom * 60.0F))) * 2.0 * Math.pow((double)f1, 1.25);
-            double d1 = Math.sin(Math.toRadians((double)(this.particleRandom * 60.0F))) * 2.0 * Math.pow((double)f1, 1.25);
+            double d0 = Math.cos(Math.toRadians(this.particleRandom * 60.0F)) * 2.0 * Math.pow(f1, 1.25);
+            double d1 = Math.sin(Math.toRadians(this.particleRandom * 60.0F)) * 2.0 * Math.pow(f1, 1.25);
             this.xd += d0 * ACCELERATION_SCALE;
             this.zd += d1 * ACCELERATION_SCALE;
-            this.yd = this.yd - (double)this.gravity;
+            this.yd = this.yd - this.gravity;
             this.rotSpeed = this.rotSpeed + this.spinAcceleration / 20.0F;
             this.oRoll = this.roll;
             this.roll = this.roll + this.rotSpeed / 20.0F;
@@ -64,22 +64,22 @@ public class CherryParticle extends TextureSheetParticle {
             }
 
             if (!this.removed) {
-                this.xd = this.xd * (double)this.friction;
-                this.yd = this.yd * (double)this.friction;
-                this.zd = this.zd * (double)this.friction;
+                this.xd = this.xd * this.friction;
+                this.yd = this.yd * this.friction;
+                this.zd = this.zd * this.friction;
             }
         }
     }
-	
+
     @OnlyIn(value = Dist.CLIENT)
 	public static class Provider implements ParticleProvider<SimpleParticleType>{
 
 		private final SpriteSet spriteSet;
-		
+
 		public Provider(SpriteSet set) {
 			this.spriteSet = set;
 		}
-		
+
 		@Override
 		public Particle createParticle(SimpleParticleType type, ClientLevel lvl, double x,
 				double y, double z, double xd, double yd, double zd) {
@@ -88,5 +88,5 @@ public class CherryParticle extends TextureSheetParticle {
 		}
 
 	}
-    
+
 }
